requirements.txt
# Core
pandas==2.2.2
requests==2.32.3
beautifulsoup4==4.12.3
tqdm==4.66.5
python-dotenv==1.0.1
langdetect==1.0.9

# Numpy pinned for PyTorch & sentence-transformers
numpy==1.26.4

# Vector search
faiss-cpu==1.8.0.post1



# LangChain ecosystem (mutually compatible)
langchain==0.2.14
langchain-community==0.2.12
langchain-text-splitters==0.2.2
langchain-mistralai==0.1.7


# Mistral SDK
mistralai==1.0.0

# Embeddings stack
transformers==4.41.2
sentence-transformers==2.7.0

# Torch CPU
torch==2.2.2

pyarrow==16.1.0
Unidecode==1.3.8
